Finished Exercise 4.21 from part4

+++ part4/blog/controllers/blogs.js
@@ -44,8 +44,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  console.log('Parameter ID', request.params.id)
+  console.log('Retrieved Blog', blog)
+  console.log('Decoded Token ID', decodedToken.id)
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response
+      .status(401)
+      .json({ error: 'Token does not correspond to Blog author' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
